requirements.txt
numpy>=1.17.4
six>=1.13.0
scipy>=1.3.2
scikit-image>=0.15.0
matplotlib>=3.1.1
ipython>=7.10.2
pillow>=6.2.1
setuptools>=42.0.1